monitoramento-veiculos-roubados/requirements.txt
easyocr
opencv-python
firebase-admin
requests
flask
flask-cors
pydantic
python-dotenv
schedule
pytest
pytest-mock